iaas-backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
marshmallow==3.22.0
python-dotenv==1.0.1
mysqlclient==2.2.4
PyMySQL==1.1.1
passlib==1.7.4
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
cryptography==42.0.8
PyMuPDF>=1.26.0,<1.28
python-docx==1.1.2
openai
sendgrid
pydantic==2.11.4
